Restructured the subdivision package
-- Integrated the doc in the header files
-- Split and moved files to have a proper internal structure and to distinguish
   between hosts, stencils and methods at the filename level.
-- Removed all instances of Polyhedron to have PolygonMesh instead
-- Cleaned off useless functions (Polyhedron_decorator remnants)
-- Improved general documentation
-- Minor performance improvements
