+++ Subdivision_method_3/doc/Subdivision_method_3/Concepts/SubdivisionMask_3.h
@@ -12,19 +12,17 @@ This concept defines types used by the mask concepts.
 class SubdivisionMask_3 {
 public:
 
-
-/// \name Types 
+/// \name Types
 /// @{
 
 /*!
 
-*/ 
+*/
   typedef unspecified_type PolygonMesh;
 
   typedef boost::graph_traits<PolygonMesh>::vertex_descriptor vertex_descriptor;
   typedef boost::graph_traits<PolygonMesh>::halfedge_descriptor halfedge_descriptor;
   typedef boost::graph_traits<PolygonMesh>::face_descriptor face_descriptor;
-  
 
 /// @}
