+++ Subdivision_method_3/doc/Subdivision_method_3/Doxyfile.in
@@ -1,6 +1,8 @@
 @INCLUDE = ${CGAL_DOC_PACKAGE_DEFAULTS}
 
 PROJECT_NAME = "CGAL ${CGAL_DOC_VERSION} - 3D Surface Subdivision Methods"
+INPUT =  ${CMAKE_SOURCE_DIR}/Subdivision_method_3/doc/Subdivision_method_3/ \ 
+         ${CMAKE_SOURCE_DIR}/Subdivision_method_3/include
 
 # custom options for this package
 EXTRACT_ALL                =  false
