+++ Subdivision_method_3/include/CGAL/Subdivision_method_3/subdivision_methods_3.h
@@ -1,5 +1,43 @@
+// ======================================================================
+//
+// Copyright (c) 2005-2017 GeometryFactory (France).  All Rights Reserved.
+//
+// This file is part of CGAL (www.cgal.org); you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public License as
+// published by the Free Software Foundation; either version 3 of the License,
+// or (at your option) any later version.
+//
+// Licensees holding a valid commercial license may use this file in
+// accordance with the commercial license agreement provided with the software.
+//
+// This file is provided AS IS with NO WARRANTY OF ANY KIND, INCLUDING THE
+// WARRANTY OF DESIGN, MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
+//
+// $URL$
+// $Id$
+//
+//
+// Author(s): Le-Jeng Shiue <Andy.Shiue@gmail.com>
+//
+// ======================================================================
+
+#ifndef CGAL_SUBDIVISION_METHODS_3_H
+#define CGAL_SUBDIVISION_METHODS_3_H
+
+#include <CGAL/basic.h>
+
+#include <vector>
+
+#include <CGAL/circulator.h>
+
+#include <CGAL/Polygon_mesh_processing/internal/named_function_params.h>
+#include <CGAL/Polygon_mesh_processing/internal/named_params_helper.h>
+
+#include <CGAL/Subdivision_method_3/subdivision_hosts_3.h>
+#include <CGAL/Subdivision_method_3/subdivision_masks_3.h>
 
 namespace CGAL {
+
 /// The namespace containing the subdivision methods.
 namespace Subdivision_method_3 {
 
@@ -10,7 +48,7 @@ A subdivision method recursively refines a coarse mesh and
 generates an ever closer approximation to a smooth surface.
 `Subdivision_method_3` consists of four subdivision methods
 and their refinement hosts. Each refinement host is a template
-function of a polyhedron class and a
+function of a polygon mesh class and a
 geometry policy class. It refines the connectivity of the
 control mesh and computes the geometry of the refined mesh.
 The geometry computation is dedicated to the custom
@@ -25,23 +63,22 @@ the primal triangle quadrisection (PTQ),
 the dual quadrilateral quadrisection (DQQ),
 and the \f$ \sqrt{3}\f$ triangulation.
 These refinements are respectively used in
-Catmull-Clark, Loop, Doo-Sabin and \f$ \sqrt{3}\f$ subdivision.
+Catmull-Clark, Loop, Doo-Sabin and \f$ \sqrt{3}\f$ subdivisions.
 
 \cgalHeading{Refinement Host}
 
 A refinement host is a template function of
-a polyhedron class and a geometry mask class. It refines
-the input polyhedron, and computes new points through
+a polygon mesh class and a geometry mask class. It refines
+the input polygon mesh, and computes new points through
 the geometry masks.
 `Subdivision_method_3` supports four refinement hosts:
 `PQQ`, `PTQ`, `DQQ` and `Sqrt3`.
 
 \image html RefSchemes.svg
 
-
 \cgalHeading{Example}
 
-This example program subdivides a polyhedral mesh with
+This example program subdivides a polygonal mesh with
 Catmull-Clark subdivision.
 
 \cgalExample{Subdivision_method_3/CatmullClark_subdivision.cpp}
@@ -53,116 +90,17 @@ Catmull-Clark subdivision.
 */
 /// @{
 
-/*!
- *
- * applies the PQQ refinement several times on the control mesh `pmesh`.
- * The geometry of the refined mesh is computed by the geometry policy `mask`.
- * This function overwrites the control mesh `pmesh` with the refined mesh.
- *
- * @tparam PolygonMesh a model of `MutableFaceGraph`
- * @tparam Mask a model of `PQQMask_3`
- * @tparam NamedParameters a sequence of \ref namedparameters
- *
- * @param pmesh a polygon mesh
- * @param mask a geometry policy mask
- * @param np optional sequence of \ref namedparameters among the ones listed below
- *
- * \cgalNamedParamsBegin
- *    \cgalParamBegin{vertex_point_map} the property map with the points associated to the vertices of `pmesh`.
- *       If this parameter is omitted, an internal property map for
- *       `CGAL::vertex_point_t` must be available in `PolygonMesh`\cgalParamEnd
- *  \cgalParamBegin{number_of_iterations} the number of subdivision steps, by default 1.
- *     \cgalParamEnd
- * \cgalNamedParamsEnd
- **/
+namespace parameters = Polygon_mesh_processing::parameters;
 
-template <class PolygonMesh, class Mask, class NamedParameters>
-void PQQ(PolygonMesh& pmesh, Mask mask, const NamedParameters& np);
+// -----------------------------------------------------------------------------
 
-/*!
- *
- * applies the PTQ refinement several times on the control mesh `pmesh`.
- * The geometry of the refined mesh is computed by the geometry policy `mask`.
- * This function overwrites the control mesh `pmesh` with the refined mesh.
- *
- * @tparam PolygonMesh a model of `MutableFaceGraph`
- * @tparam Mask a model of `PTQMask_3`
- * @tparam NamedParameters a sequence of \ref namedparameters
- *
- * @param pmesh a polygon mesh
- * @param mask a geometry policy mask
- * @param np optional sequence of \ref namedparameters among the ones listed below
- *
- * \cgalNamedParamsBegin
- *    \cgalParamBegin{vertex_point_map} the property map with the points associated to the vertices of `pmesh`.
- *       If this parameter is omitted, an internal property map for
- *       `CGAL::vertex_point_t` must be available in `PolygonMesh`\cgalParamEnd
- *  \cgalParamBegin{number_of_iterations} the number of subdivision steps, by default 1.
- *     \cgalParamEnd
- * \cgalNamedParamsEnd
- **/
-
-template <class PolygonMesh, class Mask, class NamedParameters>
-void PTQ(PolygonMesh& pmesh, Mask mask, const NamedParameters& np);
-
-/*!
- *
- * applies the DQQ refinement several times on the control mesh `pmesh`.
- * The geometry of the refined mesh is computed by the geometry policy `mask`.
- * This function overwrites the control mesh `pmesh` with the refined mesh.
- *
- * @tparam PolygonMesh a model of `MutableFaceGraph`
- * @tparam Mask a model of `DQQMask_3`
- * @tparam NamedParameters a sequence of \ref namedparameters
- *
- * @param pmesh a polygon mesh
- * @param mask a geometry policy mask
- * @param np optional sequence of \ref namedparameters among the ones listed below
- *
- * \cgalNamedParamsBegin
- *    \cgalParamBegin{vertex_point_map} the property map with the points associated to the vertices of `pmesh`.
- *       If this parameter is omitted, an internal property map for
- *       `CGAL::vertex_point_t` must be available in `PolygonMesh`\cgalParamEnd
- *  \cgalParamBegin{number_of_iterations} the number of subdivision steps, by default 1.
- *     \cgalParamEnd
- * \cgalNamedParamsEnd
- *
- * \pre `pmesh` must be a triangle mesh.
- **/
-
-template <class PolygonMesh, class Mask, class NamedParameters>
-void DQQ(PolygonMesh& pmesh, Mask mask, const NamedParameters& np);
-
-/*!
- *
- * applies the \f$ \sqrt{3}\f$ refinement several times on the control mesh `pmesh`.
- * The geometry of the refined mesh is computed by the geometry policy `mask`.
- * This function overwrites the control mesh `pmesh` with the refined mesh.
- *
- * \attention The border subdivision only happens every second subdivision step
- *            during a <em>single</em> call of this function.
- *
- * @tparam PolygonMesh a model of `MutableFaceGraph`
- * @tparam Mask a model of `Sqrt3Mask_3`
- * @tparam NamedParameters a sequence of \ref namedparameters
- *
- * @param pmesh a polygon mesh
- * @param mask a geometry policy mask
- * @param np optional sequence of \ref namedparameters among the ones listed below
- *
- * \cgalNamedParamsBegin
- *    \cgalParamBegin{vertex_point_map} the property map with the points associated to the vertices of `pmesh`.
- *       If this parameter is omitted, an internal property map for
- *       `CGAL::vertex_point_t` must be available in `PolygonMesh`\cgalParamEnd
- *  \cgalParamBegin{number_of_iterations} the number of subdivision steps, by default 1.
- *     \cgalParamEnd
- * \cgalNamedParamsEnd
- *
- * \pre `pmesh` must be a triangle mesh.
- **/
-
-template <class PolygonMesh, class Mask, class NamedParameters>
-void Sqrt3(PolygonMesh& pmesh, Mask mask, const NamedParameters& np);
+#ifndef DOXYGEN_RUNNING
+// backward compatibility
+template <class PolygonMesh>
+void CatmullClark_subdivision(PolygonMesh& pmesh, int step = 1) {
+  PQQ(pmesh, CatmullClark_mask_3<PolygonMesh>(&pmesh, get(vertex_point,pmesh)), step);
+}
+#endif
 
 /*!
  *
@@ -186,9 +124,30 @@ void Sqrt3(PolygonMesh& pmesh, Mask mask, const NamedParameters& np);
  *
  * \pre `pmesh` must be a triangle mesh.
  **/
+template <class PolygonMesh, class NamedParameters>
+void CatmullClark_subdivision(PolygonMesh& pmesh, const NamedParameters& np) {
+  using boost::choose_param;
+  using boost::get_param;
+  typedef typename GetVertexPointMap<PolygonMesh, NamedParameters>::type Vpm;
+  Vpm vpm = choose_param(get_param(np, vertex_point),
+                         get_property_map(CGAL::vertex_point, pmesh));
 
-  template <class PolygonMesh, class NamedParameters>
-void CatmullClark_subdivision(PolygonMesh& pmesh, const NamedParameters& np);
+  unsigned int step = choose_param(get_param(np, number_of_iterations), 1);
+  CatmullClark_mask_3<PolygonMesh,Vpm> mask(&pmesh, vpm);
+
+  for(unsigned int i = 0; i < step; i++)
+    internal::PQQ_1step(pmesh, vpm, mask);
+}
+
+// -----------------------------------------------------------------------------
+
+#ifndef DOXYGEN_RUNNING
+// backward compatibility
+template <class PolygonMesh>
+void Loop_subdivision(PolygonMesh& pmesh, int step = 1) {
+  PTQ(pmesh, Loop_mask_3<PolygonMesh>(&pmesh, get(vertex_point,pmesh)) , step);
+}
+#endif
 
 /*!
  *
@@ -210,9 +169,30 @@ void CatmullClark_subdivision(PolygonMesh& pmesh, const NamedParameters& np);
  *     \cgalParamEnd
  * \cgalNamedParamsEnd
  **/
+template <class PolygonMesh, class NamedParameters>
+void Loop_subdivision(PolygonMesh& pmesh, const NamedParameters& np) {
+  using boost::choose_param;
+  using boost::get_param;
+  typedef typename GetVertexPointMap<PolygonMesh, NamedParameters>::type Vpm;
+  Vpm vpm = choose_param(get_param(np, vertex_point),
+                         get_property_map(CGAL::vertex_point, pmesh));
 
-  template <class PolygonMesh, class NamedParameters>
-void Loop_subdivision(PolygonMesh& pmesh, const NamedParameters& np);
+  unsigned int step = choose_param(get_param(np, number_of_iterations), 1);
+  Loop_mask_3<PolygonMesh,Vpm> mask(&pmesh, vpm);
+
+  for(unsigned int i = 0; i < step; i++)
+    internal::PTQ_1step(pmesh, vpm, mask);
+}
+
+// -----------------------------------------------------------------------------
+
+#ifndef DOXYGEN_RUNNING
+// backward compatibility
+template <class PolygonMesh>
+void DooSabin_subdivision(PolygonMesh& pmesh, int step = 1) {
+  DQQ(pmesh, DooSabin_mask_3<PolygonMesh>(&pmesh, get(vertex_point, pmesh)), step);
+}
+#endif
 
 /*!
  *
@@ -234,9 +214,30 @@ void Loop_subdivision(PolygonMesh& pmesh, const NamedParameters& np);
  *     \cgalParamEnd
  * \cgalNamedParamsEnd
  **/
+template <class PolygonMesh, class NamedParameters>
+void DooSabin_subdivision(PolygonMesh& pmesh, const NamedParameters& np) {
+  using boost::choose_param;
+  using boost::get_param;
+  typedef typename GetVertexPointMap<PolygonMesh, NamedParameters>::type Vpm;
+  Vpm vpm = choose_param(get_param(np, vertex_point),
+                         get_property_map(CGAL::vertex_point, pmesh));
 
-  template <class PolygonMesh, class NamedParameters>
-void DooSabin_subdivision(PolygonMesh& pmesh, const NamedParameters& np);
+  unsigned int step = choose_param(get_param(np, number_of_iterations), 1);
+  DooSabin_mask_3<PolygonMesh,Vpm> mask(&pmesh, vpm);
+
+  for(unsigned int i = 0; i < step; i++)
+    internal::DQQ_1step(pmesh, vpm, mask);
+}
+
+// -----------------------------------------------------------------------------
+
+#ifndef DOXYGEN_RUNNING
+// backward compatibility
+template <class PolygonMesh>
+void Sqrt3_subdivision(PolygonMesh& pmesh, int step = 1) {
+  Sqrt3(pmesh, Sqrt3_mask_3<PolygonMesh>(&pmesh, get(vertex_point,pmesh)), step);
+}
+#endif
 
 /*!
  *
@@ -263,11 +264,25 @@ void DooSabin_subdivision(PolygonMesh& pmesh, const NamedParameters& np);
  *
  * \pre `pmesh` must be a triangle mesh.
  **/
-
 template <class PolygonMesh, class NamedParameters>
-void Sqrt3_subdivision(PolygonMesh& pmesh, const NamedParameters& np);
+void Sqrt3_subdivision(PolygonMesh& pmesh, const NamedParameters& np) {
+  using boost::choose_param;
+  using boost::get_param;
+  typedef typename GetVertexPointMap<PolygonMesh, NamedParameters>::type Vpm;
+  Vpm vpm = choose_param(get_param(np, vertex_point),
+                         get_property_map(CGAL::vertex_point, pmesh));
+
+  unsigned int step = choose_param(get_param(np, number_of_iterations), 1);
+  Sqrt3_mask_3<PolygonMesh,Vpm> mask(&pmesh, vpm);
+
+  for(unsigned int i = 0; i < step; i++)
+    internal::Sqrt3_1step(pmesh, vpm, mask, (i%2==1));
+}
 
 /// @}
 
-} /* end namespace Subdivision_method_3 */
-} /* end namespace CGAL */
+} // namespace Subdivision_method_3
+
+} // namespace CGAL
+
+#endif // CGAL_SUBDIVISION_METHODS_3_H
