+++ Subdivision_method_3/include/CGAL/subdivision_method_3.h
@@ -1,0 +1,36 @@
+// ======================================================================
+//
+// Copyright (c) 2017 GeometryFactory (France).  All Rights Reserved.
+//
+// This file is part of CGAL (www.cgal.org); you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public License as
+// published by the Free Software Foundation; either version 3 of the License,
+// or (at your option) any later version.
+//
+// Licensees holding a valid commercial license may use this file in
+// accordance with the commercial license agreement provided with the software.
+//
+// This file is provided AS IS with NO WARRANTY OF ANY KIND, INCLUDING THE
+// WARRANTY OF DESIGN, MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
+//
+// $URL$
+// $Id$
+//
+//
+// Author(s):
+//
+// ======================================================================
+
+#ifndef CGAL_SUBDIVISION_METHOD_3_H
+#define CGAL_SUBDIVISION_METHOD_3_H
+
+/**
+* \ingroup PkgSurfaceSubdivisionMethods3
+* \file CGAL/subdivision_method_3.h
+* Convenience header file including the headers for all
+* the free functions of this package.
+*/
+
+#include <CGAL/Subdivision_method_3/subdivision_methods_3.h>
+
+#endif // CGAL_SUBDIVISION_METHOD_3_H
