+++ Subdivision_method_3/package_info/Subdivision_method_3/description.txt
@@ -1,1 +1,1 @@
-Generic subdivision methods on Polyhedron_3.
+Generic subdivision methods on Polygonal Meshes.
